Pas encore fini

+++ app/Http/Controllers/TypeDocumentController.php
@@ -4,8 +4,6 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\TypeDocumentRequest;
 use App\Models\TypeDocument;
-use Illuminate\Support\Facades\Request;
-use Illuminate\Support\Facades\Storage;
 
 class TypeDocumentController extends Controller
 {
@@ -22,37 +20,6 @@ class TypeDocumentController extends Controller
     public function create(TypeDocumentRequest $request)
     {
         $request->validated();
-        // //Générer l'empreinte numérique avec la fonction MD5
-        // $empreinte = md5($request->hasFile('fichier_scanner'));
-        // //Récupérer une empreinte existant
-        // $existe = TypeDocument::where('empreinte_fichier', $empreinte)->exists();
-        // if ($existe) {
-        //     // L'empreinte du fichier existe déjà dans la base de données -> on affiche un warning.
-        //     return redirect()->route('type-document.index')->with('warning', 'Le document existe déjà dans la base de données.');
-        // } else {
-        //     $ancienFichier = "";
-        //     //On va remplacer le repertoire "storage" par "public" concernant
-        //     //l'ancien fichier du document
-        //     $ancienFichier = str_replace('storage/', 'public/', $ancienFichier);
-        //     if ($ancienFichier && Storage::exists($ancienFichier)) {
-        //         Storage::delete($ancienFichier);
-        //     }
-        //     //Stocké l'image dans la variable $path dans
-        //     //le répertoire /storage/public/upload
-        //     //et créer un lien du repertoire storage vers public
-        //     $path = $request->file('fichier_scanner')->store('public/upload');
-        //     //Rétirer le repertoire public avant de mettre dans la base de données
-        //     $replace_path = str_replace('public', '', $path);
-        //     // L'empreinte du fichier n'existe pas dans la base de données
-        //     $typeDocuments = new TypeDocument();
-        //     $typeDocuments->nom_fichier = $request->nom_fichier;
-        //     //ajouter le repertoire "storage" pour l'insertion de la base de données
-        //     $typeDocuments->fichier_scanner = "storage" . $replace_path;
-        //     $typeDocuments->empreinte_fichier = $empreinte;
-        //     $typeDocuments->save();
-        //     //Après avoir enregistrer, faire la rediretion
-        //     return redirect()->route('type-document.index')->with('status', 'Le document a bien été enregistré.');
-        // }
         
         if ($request->hasFile('fichier_scanner')) {
             // Obtenir le contenu du fichier PDF
@@ -88,21 +55,29 @@ class TypeDocumentController extends Controller
     //Delete TypeDocument
     public function delete($id)
     {
-        $typeDocument = TypeDocument::find($id); //Récupéré le document
+        $typeDocument = TypeDocument::find($id);
+
+        if (!$typeDocument) {
+            return redirect()->route('type-document.index')->with('error', 'Document non trouvé.');
+        }
 
         $typeDocument->delete();
-        //Après avoir supprimer, faire la rediretion
+
+        // Après avoir supprimé, faire la redirection avec un message de succès
         return redirect()->route('type-document.index')->with('status', 'Le document a bien été supprimé.');
     }
+
 
     public function show($id)
     {
         $document = TypeDocument::find($id);
         if ($document) {
-            return view('auth.type-document.show', compact('document'));
+            return view('auth.type-document.show', compact('document', 'id'));
         } else {
             return redirect()->route('type-document.index')->withErrors(['documentId' => 'Document non trouvé.']);
         }
     }
+
+   
 
 }
